feat: remove `--no-docker` flag + split into submodules
